pyproject.toml
[project]
name = "hemapex"
version = "0.1.0"
description = "Unlock LLM APIs power to retrieve informations such as treatments and relapses in patients electronic medical records"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.51.0",
    "logging>=0.4.9.6",
    "mistralai>=1.9.11",
    "openai>=2.7.2",
    "pandas>=2.3.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "pre-commit>=4.4.0",
    "ruff>=0.14.4",
    "streamlit>=1.51.0",
]
